+++ .psalm/baseline.xml
@@ -818,9 +818,6 @@
             $configuration->colors(),
             $configuration->columns(),
             $configuration->source(),
-            $configuration->restrictDeprecations(),
-            $configuration->restrictNotices(),
-            $configuration->restrictWarnings(),
         )]]></code>
       <code><![CDATA[new DefaultProgressPrinter(
             self::$printer,
@@ -828,9 +825,6 @@
             $configuration->colors(),
             $configuration->columns(),
             $configuration->source(),
-            $configuration->restrictDeprecations(),
-            $configuration->restrictNotices(),
-            $configuration->restrictWarnings(),
         )]]></code>
     </MissingThrowsDocblock>
     <NullableReturnStatement>
