+++ ChangeLog-10.2.md
@@ -4,4 +4,10 @@ All notable changes of the PHPUnit 10.2 release series are documented in this fi
 
 ## [10.2.0] - 2023-06-02
 
+### Deprecated
+
+* `PHPUnit\TextUI\Configuration\Configuration::restrictDeprecations()` (use `source()->restrictDeprecations()` instead)
+* `PHPUnit\TextUI\Configuration\Configuration::restrictNotices()` (use `source()->restrictNotices()` instead)
+* `PHPUnit\TextUI\Configuration\Configuration::restrictWarnings()` (use `source()->restrictWarnings()` instead)
+
 [10.2.0]: https://github.com/sebastianbergmann/phpunit/compare/10.1...main
