+++ src/Runner/TestResult/Collector.php
@@ -50,16 +50,6 @@ use PHPUnit\TextUI\Configuration\SourceFilter;
 final class Collector
 {
     private readonly Source $source;
-    private readonly bool $restrictDeprecations;
-    private readonly bool $restrictNotices;
-    private readonly bool $restrictWarnings;
-    private readonly bool $ignoreSuppressionOfDeprecations;
-    private readonly bool $ignoreSuppressionOfPhpDeprecations;
-    private readonly bool $ignoreSuppressionOfErrors;
-    private readonly bool $ignoreSuppressionOfNotices;
-    private readonly bool $ignoreSuppressionOfPhpNotices;
-    private readonly bool $ignoreSuppressionOfWarnings;
-    private readonly bool $ignoreSuppressionOfPhpWarnings;
     private int $numberOfTests                       = 0;
     private int $numberOfTestsRun                    = 0;
     private int $numberOfAssertions                  = 0;
@@ -160,7 +150,7 @@ final class Collector
      * @throws EventFacadeIsSealedException
      * @throws UnknownSubscriberTypeException
      */
-    public function __construct(Facade $facade, Source $source, bool $restrictDeprecations, bool $restrictNotices, bool $restrictWarnings, bool $ignoreSuppressionOfDeprecations, bool $ignoreSuppressionOfPhpDeprecations, bool $ignoreSuppressionOfErrors, bool $ignoreSuppressionOfNotices, bool $ignoreSuppressionOfPhpNotices, bool $ignoreSuppressionOfWarnings, bool $ignoreSuppressionOfPhpWarnings)
+    public function __construct(Facade $facade, Source $source)
     {
         $facade->registerSubscribers(
             new ExecutionStartedSubscriber($this),
@@ -189,17 +179,7 @@ final class Collector
             new TestRunnerTriggeredWarningSubscriber($this),
         );
 
-        $this->source                             = $source;
-        $this->restrictDeprecations               = $restrictDeprecations;
-        $this->restrictNotices                    = $restrictNotices;
-        $this->restrictWarnings                   = $restrictWarnings;
-        $this->ignoreSuppressionOfDeprecations    = $ignoreSuppressionOfDeprecations;
-        $this->ignoreSuppressionOfPhpDeprecations = $ignoreSuppressionOfPhpDeprecations;
-        $this->ignoreSuppressionOfErrors          = $ignoreSuppressionOfErrors;
-        $this->ignoreSuppressionOfNotices         = $ignoreSuppressionOfNotices;
-        $this->ignoreSuppressionOfPhpNotices      = $ignoreSuppressionOfPhpNotices;
-        $this->ignoreSuppressionOfWarnings        = $ignoreSuppressionOfWarnings;
-        $this->ignoreSuppressionOfPhpWarnings     = $ignoreSuppressionOfPhpWarnings;
+        $this->source = $source;
     }
 
     public function result(): TestResult
@@ -384,11 +364,11 @@ final class Collector
 
     public function testTriggeredDeprecation(DeprecationTriggered $event): void
     {
-        if (!$this->ignoreSuppressionOfDeprecations && $event->wasSuppressed()) {
+        if (!$this->source->ignoreSuppressionOfDeprecations() && $event->wasSuppressed()) {
             return;
         }
 
-        if ($this->restrictDeprecations && !(new SourceFilter)->includes($this->source, $event->file())) {
+        if ($this->source->restrictDeprecations() && !(new SourceFilter)->includes($this->source, $event->file())) {
             return;
         }
 
@@ -401,11 +381,11 @@ final class Collector
 
     public function testTriggeredPhpDeprecation(PhpDeprecationTriggered $event): void
     {
-        if (!$this->ignoreSuppressionOfPhpDeprecations && $event->wasSuppressed()) {
+        if (!$this->source->ignoreSuppressionOfPhpDeprecations() && $event->wasSuppressed()) {
             return;
         }
 
-        if ($this->restrictDeprecations && !(new SourceFilter)->includes($this->source, $event->file())) {
+        if ($this->source->restrictDeprecations() && !(new SourceFilter)->includes($this->source, $event->file())) {
             return;
         }
 
@@ -427,7 +407,7 @@ final class Collector
 
     public function testTriggeredError(ErrorTriggered $event): void
     {
-        if (!$this->ignoreSuppressionOfErrors && $event->wasSuppressed()) {
+        if (!$this->source->ignoreSuppressionOfErrors() && $event->wasSuppressed()) {
             return;
         }
 
@@ -440,11 +420,11 @@ final class Collector
 
     public function testTriggeredNotice(NoticeTriggered $event): void
     {
-        if (!$this->ignoreSuppressionOfNotices && $event->wasSuppressed()) {
+        if (!$this->source->ignoreSuppressionOfNotices() && $event->wasSuppressed()) {
             return;
         }
 
-        if ($this->restrictNotices && !(new SourceFilter)->includes($this->source, $event->file())) {
+        if ($this->source->restrictNotices() && !(new SourceFilter)->includes($this->source, $event->file())) {
             return;
         }
 
@@ -457,11 +437,11 @@ final class Collector
 
     public function testTriggeredPhpNotice(PhpNoticeTriggered $event): void
     {
-        if (!$this->ignoreSuppressionOfPhpNotices && $event->wasSuppressed()) {
+        if (!$this->source->ignoreSuppressionOfPhpNotices() && $event->wasSuppressed()) {
             return;
         }
 
-        if ($this->restrictNotices && !(new SourceFilter)->includes($this->source, $event->file())) {
+        if ($this->source->restrictNotices() && !(new SourceFilter)->includes($this->source, $event->file())) {
             return;
         }
 
@@ -474,11 +454,11 @@ final class Collector
 
     public function testTriggeredWarning(WarningTriggered $event): void
     {
-        if (!$this->ignoreSuppressionOfWarnings && $event->wasSuppressed()) {
+        if (!$this->source->ignoreSuppressionOfWarnings() && $event->wasSuppressed()) {
             return;
         }
 
-        if ($this->restrictWarnings && !(new SourceFilter)->includes($this->source, $event->file())) {
+        if ($this->source->restrictWarnings() && !(new SourceFilter)->includes($this->source, $event->file())) {
             return;
         }
 
@@ -491,11 +471,11 @@ final class Collector
 
     public function testTriggeredPhpWarning(PhpWarningTriggered $event): void
     {
-        if (!$this->ignoreSuppressionOfPhpWarnings && $event->wasSuppressed()) {
+        if (!$this->source->ignoreSuppressionOfPhpWarnings() && $event->wasSuppressed()) {
             return;
         }
 
-        if ($this->restrictWarnings && !(new SourceFilter)->includes($this->source, $event->file())) {
+        if ($this->source->restrictWarnings() && !(new SourceFilter)->includes($this->source, $event->file())) {
             return;
         }
