+++ src/TextUI/Output/Facade.php
@@ -159,16 +159,6 @@ final class Facade
             $configuration->colors(),
             $configuration->columns(),
             $configuration->source(),
-            $configuration->restrictDeprecations(),
-            $configuration->restrictNotices(),
-            $configuration->restrictWarnings(),
-            $configuration->ignoreSuppressionOfDeprecations(),
-            $configuration->ignoreSuppressionOfPhpDeprecations(),
-            $configuration->ignoreSuppressionOfErrors(),
-            $configuration->ignoreSuppressionOfNotices(),
-            $configuration->ignoreSuppressionOfPhpNotices(),
-            $configuration->ignoreSuppressionOfWarnings(),
-            $configuration->ignoreSuppressionOfPhpWarnings(),
         );
 
         self::$defaultProgressPrinter = true;
