+++ tests/unit/TextUI/SourceFilterTest.php
@@ -36,6 +36,16 @@ final class SourceFilterTest extends TestCase
                     ),
                     FilterDirectoryCollection::fromArray([]),
                     FileCollection::fromArray([]),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using file, but excluded using directory' => [
@@ -58,6 +68,16 @@ final class SourceFilterTest extends TestCase
                         ]
                     ),
                     FileCollection::fromArray([]),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using file, but excluded using file' => [
@@ -76,6 +96,16 @@ final class SourceFilterTest extends TestCase
                             new File($fixtureDirectory . '/a/PrefixSuffix.php'),
                         ]
                     ),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using directory' => [
@@ -94,6 +124,16 @@ final class SourceFilterTest extends TestCase
                     FileCollection::fromArray([]),
                     FilterDirectoryCollection::fromArray([]),
                     FileCollection::fromArray([]),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using directory, but excluded using file' => [
@@ -116,6 +156,16 @@ final class SourceFilterTest extends TestCase
                             new File($fixtureDirectory . '/a/PrefixSuffix.php'),
                         ]
                     ),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using directory, but excluded using directory' => [
@@ -142,6 +192,16 @@ final class SourceFilterTest extends TestCase
                         ]
                     ),
                     FileCollection::fromArray([]),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
         ];
