+++ tests/unit/TextUI/SourceMapperTest.php
@@ -37,6 +37,16 @@ final class SourceMapperTest extends TestCase
                     ),
                     FilterDirectoryCollection::fromArray([]),
                     FileCollection::fromArray([]),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using file, but excluded using directory' => [
@@ -59,6 +69,16 @@ final class SourceMapperTest extends TestCase
                         ]
                     ),
                     FileCollection::fromArray([]),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using file, but excluded using file' => [
@@ -77,6 +97,16 @@ final class SourceMapperTest extends TestCase
                             new File($fixtureDirectory . '/a/PrefixSuffix.php'),
                         ]
                     ),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using directory' => [
@@ -107,6 +137,16 @@ final class SourceMapperTest extends TestCase
                     FileCollection::fromArray([]),
                     FilterDirectoryCollection::fromArray([]),
                     FileCollection::fromArray([]),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using directory, but excluded using file' => [
@@ -140,6 +180,16 @@ final class SourceMapperTest extends TestCase
                             new File($fixtureDirectory . '/a/PrefixSuffix.php'),
                         ]
                     ),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
             'file included using directory, but excluded using directory' => [
@@ -171,6 +221,16 @@ final class SourceMapperTest extends TestCase
                         ]
                     ),
                     FileCollection::fromArray([]),
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
                 ),
             ],
         ];
